HSV should use Doubles, not Ints
Ignore-this: 161accb1419b0303c76593bbeb12c33c
Spotted by Michael deLorimier

darcs-hash:20090908121105-97e08-f498b8e934f78137dfe12d74271bd3246d731f9a

+++ Data/GraphViz/Attributes.hs
@@ -1121,9 +1121,9 @@ data Color = RGB { red   :: Word8
                   , blue  :: Word8
                   , alpha :: Word8
                   }
-           | HSV { hue        :: Int
-                 , saturation :: Int
-                 , value      :: Int
+           | HSV { hue        :: Double
+                 , saturation :: Double
+                 , value      :: Double
                  }
            | ColorName String
              deriving (Eq, Show, Read)
